+++ release-notes-3.7.0.md
@@ -3,6 +3,9 @@
 #### Version Number
 ${version-number}
 
+#### Breaking change
+- US975260: Refactor to remove HPE references in package names
+
 #### New Features
 
 #### Known Issues
