+++ worker-batch-testing/src/main/java/com/github/jobservice/worker/batch/BatchTestExpectation.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2016-2024 Open Text.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.github.jobservice.worker.batch;
+
+import com.github.workerframework.worker.api.TaskMessage;
+
+import java.util.List;
+
+public class BatchTestExpectation {
+    private List<TaskMessage> subTasks;
+
+    public List<TaskMessage> getSubTasks() {
+        return subTasks;
+    }
+
+    public void setSubTasks(List<TaskMessage> subTasks) {
+        this.subTasks = subTasks;
+    }
+}
