+++ worker-batch-testing/src/main/resources/META-INF/services/com.github.workerframework.worker.testing.execution.TestControllerProvider
@@ -1,0 +1,1 @@
+com.github.jobservice.worker.batch.BatchTestControllerProvider
