--- worker-batch-testing/src/main/resources/META-INF/services/com.hpe.caf.worker.testing.execution.TestControllerProvider
@@ -1,1 +1,0 @@
-com.hpe.caf.worker.batch.BatchTestControllerProvider
