+++ worker-batch-testing/src/test/java/CompareTaskDataAsDocumentWorkerTaskTest.java
@@ -19,7 +19,7 @@ import com.github.workerframework.worker.api.DataStore;
 import com.github.workerframework.worker.api.DataStoreException;
 import com.github.workerframework.worker.api.TaskMessage;
 import com.github.workerframework.worker.api.TaskStatus;
-import com.hpe.caf.worker.batch.BatchResultValidationProcessor;
+import com.github.jobservice.worker.batch.BatchResultValidationProcessor;
 import static org.testng.Assert.fail;
 import static org.testng.Assert.assertTrue;
 import static org.testng.Assert.assertFalse;
