+++ worker-batch/src/main/java/com/github/jobservice/worker/batch/BatchWorkerFactoryProvider.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2016-2024 Open Text.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.github.jobservice.worker.batch;
+
+import com.github.cafapi.common.api.Codec;
+import com.github.cafapi.common.api.ConfigurationSource;
+import com.github.workerframework.worker.api.DataStore;
+import com.github.workerframework.worker.api.WorkerException;
+import com.github.workerframework.worker.api.WorkerFactory;
+import com.github.workerframework.worker.api.WorkerFactoryProvider;
+
+public class BatchWorkerFactoryProvider implements WorkerFactoryProvider {
+
+    @Override
+    public WorkerFactory getWorkerFactory(ConfigurationSource configSource, DataStore dataStore, Codec codec) throws WorkerException {
+        return new BatchWorkerFactory(configSource,dataStore,codec,BatchWorkerConfiguration.class, BatchWorkerTask.class);
+    }
+}
