+++ worker-batch/src/main/resources/META-INF/services/com.github.workerframework.worker.api.WorkerFactoryProvider
@@ -1,0 +1,1 @@
+com.github.jobservice.worker.batch.BatchWorkerFactoryProvider
