--- worker-batch/src/main/resources/META-INF/services/com.hpe.caf.api.worker.WorkerFactoryProvider
@@ -1,1 +1,0 @@
-com.hpe.caf.worker.batch.BatchWorkerFactoryProvider
