+++ worker-batch/src/test/java/BatchWorkerServicesTest.java
@@ -16,6 +16,13 @@
 import com.github.cafapi.common.api.Codec;
 import com.github.cafapi.common.api.CodecException;
 import com.github.cafapi.common.codecs.jsonlzf.JsonCodec;
+import com.github.jobservice.worker.batch.BatchWorker;
+import com.github.jobservice.worker.batch.BatchWorkerConfiguration;
+import com.github.jobservice.worker.batch.BatchWorkerConstants;
+import com.github.jobservice.worker.batch.BatchWorkerPlugin;
+import com.github.jobservice.worker.batch.BatchWorkerServicesImpl;
+import com.github.jobservice.worker.batch.BatchWorkerTask;
+import com.github.jobservice.worker.batch.ReturnValueBehaviour;
 import com.github.workerframework.worker.api.DataStore;
 import com.github.workerframework.worker.api.TaskFailedException;
 import com.github.workerframework.worker.api.TaskStatus;
@@ -30,7 +37,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.io.TempDir;
-import org.mockito.Mock;
 import org.mockito.Mockito;
 
 import java.io.File;
